Fix bug where servers rejected event stream content type (#4322)
## Motivation and Context
There is a bug in the generated server SDK where for operations with an
output event stream, it rejects `application/vnd.amazon.eventstream`,
instead expecting `application/cbor`.

This is incorrect. However, we need to maintain backwards compatibility
with existing clients. This adds a new fallback option that protocols
can set to handle this for RpcV2Cbor.

## Description
- Small refactoring to extract a mime-type generator
- Allow accepting multiple headers as required.

## Testing
I briefly went down the path of trying to get this to work as a protocol
test but ultimately it became too much of a yak shave.

I added a unit test that should be sufficient.

## Checklist
<!--- If a checkbox below is not applicable, then please DELETE it
rather than leaving it unchecked -->
- [ ] For changes to the smithy-rs codegen or runtime crates, I have
created a changelog entry Markdown file in the `.changelog` directory,
specifying "client," "server," or both in the `applies_to` key.
- [ ] For changes to the AWS SDK, generated SDK code, or SDK runtime
crates, I have created a changelog entry Markdown file in the
`.changelog` directory, specifying "aws-sdk-rust" in the `applies_to`
key.

----

_By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice._