Bump Smithty to 1.62 (#4323)
## Motivation and Context
Upgrades Smithy to
[1.62.0](https://github.com/smithy-lang/smithy/releases/tag/1.62.0)

## Description
With Smithy 1.62.0, we need to use the updated partitions file. As a
result, we updated service models in this repository (copied them over
from `aws-sdk-rust/aws-models`).

Two unicode substring tests failed upstream, within the Smithy team
itself, as part of 1.62.0. The team deleted them with the intention that
they will add new ones to the public repo, so we manually replicated the
deletion in `substring.smithy` (maybe we don't have to keep a local copy
of `codegen-client/src/test/resources/META-INF/smithy/endpoint-tests`
once tests are added to public Smithy repo).

## Testing
- CI
- Verified build and running tests for dynamodb for today's failed
release.

We should be able to ignore semver compliance check since it's comparing
the ec2 crate from the old model against that from the updated model;
the one from the updated model is already out in the wild on crates.io

----

_By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice._