Add server support for initial-request messages
This adds support for `initial-request` event stream messages on the server and fixes #4325.

This support is complete (in that it handles both empty and non-empty initial requests and is
backwards compatible with clients that do not send initial requests when not required.

BUT: it is not complete in that the errors it sends when the initial request are invalid will not easily make it back to clients. (#4343)

The error needs to be sent over the event-stream, but the current design of the traits in aws-smithy-http-server do not allow opening the event stream in this context.