Bump aws-smithy-http to 0.63 per request from server team (#4334)
## Motivation and Context
<!--- Why is this change required? What problem does it solve? -->
<!--- If it fixes an open issue, please link to the issue here -->
Request from @drganjoo 

>Also, in feature/http-1.x could we please bump version of
aws-smithy-http to 0.63.x instead of
[0.62.4](https://github.com/smithy-lang/smithy-rs/blob/feature/http-1.x/rust-runtime/aws-smithy-http/Cargo.toml#L3).
This way on the server side we can keep on relying on 0.62.x to have
http/0.x support for a while. Thanks.

Note: I merged main back into the base `feature/http-1.x` branch as part
of making this change

## Testing
<!--- Please describe in detail how you tested your changes -->
<!--- Include details of your testing environment, and the tests you ran
to -->
<!--- see how your change affects other areas of the code, etc. -->

Failing tests are all server related and expected since the server side
of the codebase hasn't done this upgrade yet.

----

_By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice._