Add ability to write `integration-tests` for servers (#4332)
## Motivation and Context

This `integration-tests` directory makes it possible to write
Rust code that you can run directly against the generated servers
from the protocol test models. This makes it easier to test things
like event streams behavior and other specifics that are hard to
test with protocol tests alone.

## Description

I added a test for #4325 (showing the broken behavior) that we can use
when we fix this issue. This will also make it easy to test a number of
other open issues that have been found.

To make this easy I also added a `EventStreamSender::once` method for
convenience. The `ManualEventStream` sender could be upstreamed at a
later date so it is more broadly usable but I didn't want to pin down a
public API for it right now.
<!--- Describe your changes in detail -->

----

_By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice._