Add server support for initial-request messages (#4344)
## Motivation and Context
#4325: Properly handle initial requests for servers
Fixes: #4235 

## Description
<!--- Describe your changes in detail -->
This adds support for `initial-request` event stream messages on the
server and fixes #4325.

This support is complete (in that it handles both empty and non-empty
initial requests and is backwards compatible with clients that do not
send initial requests when not required.

BUT: it is not complete in that the errors it sends when the initial
request are invalid will not easily make it back to clients. (#4343)

The error needs to be sent over the event-stream, but the current design
of the traits in aws-smithy-http-server do not allow opening the event
stream in this context.

## Testing
Built on the test suite created earlier to test all different cases of
event stream error handling.

## Checklist
<!--- If a checkbox below is not applicable, then please DELETE it
rather than leaving it unchecked -->
- [x] For changes to the smithy-rs codegen or runtime crates, I have
created a changelog entry Markdown file in the `.changelog` directory,
specifying "client," "server," or both in the `applies_to` key.
- [x] For changes to the AWS SDK, generated SDK code, or SDK runtime
crates, I have created a changelog entry Markdown file in the
`.changelog` directory, specifying "aws-sdk-rust" in the `applies_to`
key.

----

_By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice._