Fixing bug with enums as eventHeaders (#4337)
## Motivation and Context
<!--- Why is this change required? What problem does it solve? -->
<!--- If it fixes an open issue, please link to the issue here -->
Fixing a bug that was discovered with smithy `enum` and `intEnum` shapes
that are marked as `@eventHeader`. This came up in an internal
investigation around a failed preview build.

## Description
<!--- Describe your changes in detail -->
Added handling of `EnumShape` in `EventStreamUnmarshallerGenerator` and
`EventStreamMarshallerGenerator`.

## Testing
<!--- Please describe in detail how you tested your changes -->
<!--- Include details of your testing environment, and the tests you ran
to -->
<!--- see how your change affects other areas of the code, etc. -->
Added new `enum` and `intEnum` shapes to the evenstream test model. They
are only included in client tests since the server throws a validation
error on enums in event streams (see
https://github.com/smithy-lang/smithy/issues/1388 for more info).

Also tested against the model from the failed preview build and a `cargo
test --all-features` was successful.


----

_By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice._