Allow clients to always accept intial response
Unexpected and contrary to documentation, clients do not currently always accept initial responses. This fixes that issue and updates clients to
always accept initial responses for event streams.

Because of this, we must also change the default to `false` for the setting to avoid backwards compatibility issues with existing clients.