Add support for `initial-response` event streams on servers (#4352)
## Motivation and Context

[Event Stream](https://smithy.io/2.0/spec/streaming.html#event-streams)
operations over RPC protocols (that do not rely on HTTP headers), use
`initial-request` and `initial-response` objects to send data fields
that are not bound to the event stream itself.

However, there are two major bug in the current implementation:
- #4353: Clients error if an initial-response is sent when one is not
expected
- #4345: Servers do not send initial responses (ever)

## Description
This adds two things:
1. Optional (and disabled by default) sending of an initial-response
message in all cases (even when the resulting message would be empty).
Servers that communicate with `smithy-java` clients MUST enable this
setting currently. This is disabled by default because clients in the
wild today do not accept an empty `initial-response`.
3. Support for sending non-empty initial response messages

## Testing
- Extensive unit testing

## Checklist
<!--- If a checkbox below is not applicable, then please DELETE it
rather than leaving it unchecked -->
- [x] For changes to the smithy-rs codegen or runtime crates, I have
created a changelog entry Markdown file in the `.changelog` directory,
specifying "client," "server," or both in the `applies_to` key.
- [x] For changes to the AWS SDK, generated SDK code, or SDK runtime
crates, I have created a changelog entry Markdown file in the
`.changelog` directory, specifying "aws-sdk-rust" in the `applies_to`
key.

----

_By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice._