include the protocol name in crate metadata (#4357)
include it as `package.metadata.smithy.protocol`

this allows to easily tell what protocol was used for a given crate

example protocol name: `aws.protocols#restJson1`

## Motivation and Context

Currently, when seeing a Smithy-generated crate for a customer, it can
be annoying to find out which protocol that crate was generated for.
Include the metadata in the Cargo.toml to make that easier.

cc @drganjoo you wanted this originally.

## Description

Include the protocol name in `package.metadata.smithy.protocol` in
`Cargo.toml`

## Testing

Added the check to existing unit tests

Not sure whether this is the best place to put them (asking @drganjoo)

## Checklist
<!--- If a checkbox below is not applicable, then please DELETE it
rather than leaving it unchecked -->
- [x] For changes to the smithy-rs codegen or runtime crates, I have
created a changelog entry Markdown file in the `.changelog` directory,
specifying "client," "server," or both in the `applies_to` key.

----

_By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice._

Co-authored-by: Ariel Ben-Yehuda <arielby@amazon.com>
Co-authored-by: Landon James <lnj@amazon.com>