Bump MSRV to 1.88.0 (#4367)
## Motivation and Context
Bumps MSRV to 1.88.0

## Description
Although the number of files changed is large, the changes themselves
are straightforward. During MSRV upgrade to 1.88.0, one of the clippy
warnings enforced is `clippy::uninlined_format_args`, i.e. suggesting to
update
```
format!("{}", a);
```
to
```
format!("{a}");
```
The PR swept through the codebase and updated almost all the
occurrences, except for:
- codegen-client (suppressed the following)
-
[RequestBindingGenerator](https://github.com/smithy-lang/smithy-rs/pull/4367/files#diff-fea7e7ebc2f668be75ee8e15dd463bd38064b416eea4c0d272e17ce52e2e7598R138-R142)
-
[‎EndpointTraitBindingGenerator](https://github.com/smithy-lang/smithy-rs/pull/4367/files#diff-bf24c9ba358f747e62b6603c3ddfae91340a9971c41b268a148f705dfface9deR92-R96)
- codegen-server
  - https://github.com/smithy-lang/smithy-rs/issues/4366

## Testing
- CI

## Checklist
<!--- If a checkbox below is not applicable, then please DELETE it
rather than leaving it unchecked -->
- [x] For changes to the smithy-rs codegen or runtime crates, I have
created a changelog entry Markdown file in the `.changelog` directory,
specifying "client," "server," or both in the `applies_to` key.
- [x] For changes to the AWS SDK, generated SDK code, or SDK runtime
crates, I have created a changelog entry Markdown file in the
`.changelog` directory, specifying "aws-sdk-rust" in the `applies_to`
key.

----

_By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice._