generate and check rust-version for Smithy packages
Previously to this CR, many Smithy packages did not declare their MSRV.
This causes problems because the Cargo MSRV-aware resolver is by design
not fully backtracking.

For example, the current version of aws-config (1.8.10) depends on
aws-sdk-sts 1.90.0, which has a pretty high MSRV (1.88). Because of how
the MSRV-aware resolver works, as long as aws-config does not declare an
MSRV, everyone using rustc < 1.88 with an `aws-config = "1"` dependency
will pick aws-config 1.8.10 (because the resolver is eager), and then
fail to compile because all aws-sdk-sts versions that work with 1.8.10
don't work with rustc < 1.88.

This means that everyone using Smithy and updating their Cargo.lock had
to use rustc 1.88, which is quite annoying churn, especially for
libraries that want to have a fixed MSRV.

If aws-config 1.8.10 had a correctly-configured MSRV, then for example
cargo using rustc 1.87 would pick aws-config 1.8.9, which can use
aws-sdk-sts 1.89, which would keep working.

I don't think we should yank aws-config 1.8.10 as all users have already
updated their rustc, but this will prevent this problem from happening
during the next MSRV bump.