generate and check rust-version for Smithy packages (#4414)
## Motivation and Context
<!--- Why is this change required? What problem does it solve? -->
<!--- If it fixes an open issue, please link to the issue here -->

Previously to this CR, many Smithy packages, including
<https://crates.io/crates/aws-config>, did not declare their MSRV. This
causes problems because the Cargo MSRV-aware resolver is by design not
fully backtracking.

For example, the current version of aws-config (1.8.10) depends on
aws-sdk-sts 1.90.0, which has a pretty high MSRV (1.88). Because of how
the MSRV-aware resolver works, as long as aws-config does not declare an
MSRV, everyone using rustc < 1.88 with an `aws-config = "1"` dependency
will pick aws-config 1.8.10 (because the resolver is eager), and then
fail to compile because all aws-sdk-sts versions that work with 1.8.10
don't work with rustc < 1.88.

This means that everyone using Smithy and updating their Cargo.lock had
to use rustc 1.88, which is quite annoying churn, especially for
libraries that want to have a fixed MSRV.

If aws-config 1.8.10 had a correctly-configured MSRV, then for example
cargo using rustc 1.87 would pick aws-config 1.8.9, which can use
aws-sdk-sts 1.89, which would keep working.

I don't think we should yank aws-config 1.8.10 as all users have already
updated their rustc, but this will prevent this problem from happening
during the next MSRV bump.

## Description
<!--- Describe your changes in detail -->

Added a script that uses `cargo msrv` to verify MSRVs and changed MSRVs
to obey it.

## Testing
<!--- Please describe in detail how you tested your changes -->
<!--- Include details of your testing environment, and the tests you ran
to -->
<!--- see how your change affects other areas of the code, etc. -->

This does not change any code. The MSRV is tested by the CI script.

## Checklist
<!--- If a checkbox below is not applicable, then please DELETE it
rather than leaving it unchecked -->
- [x] For changes to the smithy-rs codegen or runtime crates, I have
created a changelog entry Markdown file in the `.changelog` directory,
specifying "client," "server," or both in the `applies_to` key.
- [x] For changes to the AWS SDK, generated SDK code, or SDK runtime
crates, I have created a changelog entry Markdown file in the
`.changelog` directory, specifying "aws-sdk-rust" in the `applies_to`
key.

----

_By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice._

---------

Co-authored-by: Ariel Ben-Yehuda <arielby@amazon.com>
Co-authored-by: Landon James <lnj@amazon.com>