Add token bucket success rewards, time based refill, fractional tokens (#4450)
## Motivation and Context
Provides the ability to grant tokens as a reward for successful
requests, and the ability to grant time-based refill of tokens. This
provides more granular controls over the Token Bucket for users to
implement their own token strategy when needed.

## Description
Added the refill capabilities and all needed builders

## Testing
Added the following fields to the TokenBucket. They are required to
allot rewards and track the passing of time. Fractional tokens is
required so that success rewards and time based refill can be fractional
values.
```
    success_reward: f32,
    fractional_tokens: Arc<AtomicF32>,
    refill_rate: f32,
    time_source: SharedTimeSource,
    creation_time: SystemTime,
    last_refill_age_secs: Arc<AtomicU32>,
```

Added setters for:
```
    success_reward: f32, -> defaults to 0.0
    refill_rate: f32, -> defaults to 0.0
```

When a request succeeds, the tokens for success are calculated and added
to fractional_tokens.
When a request attempts to acquire a token from the token bucket in
order to make a call, the time based refill is calculated and added to
fractional_tokens. Once that is done, the fractional tokens are
converted to whole tokens, and used to add more permits to the
TokenBucket. At this point, we will either allow a retry request if
there are sufficient tokens, or deny if there are not.

In cases where success reward or refill rate are 0, we do not perform
these evaluations.

For time based refill, we use a compare_exchange to ensure only one
thread updates the time based refill, to prevent overgeneration of
rewards.


This was tested by adding unit tests and ensuring old tests pass. I also
tested this in the code base where I intend to consume this feature and
validated that the behavior was as required and expected.

For existing users of the code, no change in behavior should be
observed. All defaults are set such that they will result in the
behavior present before this change.

## Checklist
- [X] For changes to the smithy-rs codegen or runtime crates, I have
created a changelog entry Markdown file in the `.changelog` directory,
specifying "client," "server," or both in the `applies_to` key.

----

_By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice._